requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
chromadb==0.6.3
sentence-transformers==3.3.1
PyMuPDF==1.25.3
openpyxl==3.1.5
python-pptx==1.0.2
httpx==0.28.1
python-dotenv==1.0.1
